fix: missing reference directives in outputted declarations
I added a lint task which uncovered that the test plugin lacked types
for `@types/node/path` which `pathe` relies on. This is mostly an issue
because `pathe` relies this type package, but has it marked as a devDep.

I thought I had fixed this with the reference directive, but turns out
tsc strips that from the .d.ts in the declaration file. They're still
in the built js file, which is how I assume it was somehow semi-working.

The fix uses the new `preserve` attribute for triple-slash directives
added in typescript 5.5 which is still in beta.